Change base to tidyverse image
It turns out that the r-base image is based on debian unstable,
and it's always just the latest version of R available, so it's
not really meant to be the basic R image for people to use.

Rather than switching to r-ver, which is the basic R image, I'm
trying out the tidyverse image, because installing R packages
takes forever, and getting all the tidyverse packages in the base
image adds extra memory (due to unused packages in the final image),
but saves considerable time when building, which has always been
the main pain point.