Run selenium in separate docker container
Google Cloud Run is being a dick about running selenium in the
same container as the application. My theory is that selenium
is using up too many resources (memory, cpu), and GCR is killing
the selenium server. So, I'm trying to run selenium in a separate
container per the recommendation in the RSelenium documentation.