Wait for Splash & add directory to check
I kept getting test failures when runing 'devtools::check' even
though they passed with 'devtools::test'. It seems to have
something to do with how 'here' builds the filepath when running
'check' causing errors when we try to read the Lua file.

I don't know why defining the 'check_dir' fixes this, but I'm
sick of tinkering with it and am just going to take the win.