Remove Tab webscraper as backup betting data source
Scraping Tab works on local, but not in Google Cloud Run, which
suggests that they might block some IP addresses associated with
cloud services to prevent this very activity. I might be able to
get it working with a Splash instance hosted on DigitalOcean or
some such, but I don't feel like paying $5 a month for something
I only need in the case of an AFL season played in the midst of
a global pandemic, and even then, only like 3 times.